fruid-util: Hide absent FRU errors when displaying all
Summary:
While it should be an error when user wants to view the FRUID
of an absent FRU, it should not be an error when the user
wants to display FRUID from all FRUs and one or more FRUs
are absent. In a marginally complicated system this is not
a failure (All cards need not be present in all configurations!).
This changes the code to make sure we do not print an error
and fail the call if the user is printing all FRUIDs if
the FRU happens to be absent.

Also fix the issue where utility would not exit with error if
retrieving FRUID fails for any of the present FRUs.

Test Plan:
Build and test on tiogapass.
1. Test if printing absent fru returns failure
2. Test if printing all does not cause failure due to absent
(Note in FBTP NIC wll still cause failure due to lack of support)

Reviewed By: benwei13

fbshipit-source-id: 23cee0a612