Fix hostname regression in base_fw_upgrade_test.py (#144)
Summary:
A recent change in base_fw_upgrade_test.py makes a condition always evaluate to True, the result being that ".fbinfra.net" is always appended to the BMC hostname. This broke the test for our Elbert testing.

## Testing
Verified that the Elbert fw_upgrade_test is now working correctly and that the hostname is correctly set.

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/144

Reviewed By: tao-ren

fbshipit-source-id: 12d42bf840