Ensure oob-eeprom-util.sh always writes to stderr when failing (#143)
Summary:
Occasionally, the test_oob_eeprom CIT test fails when run by
automated tests. When it fails, the error message from the CIT test simply
indicates that the oob-eeprom-util.sh read command has a non-zero exit
code. This implies that nothing was written to stderr because if there had,
then this would have been present in the error message.

In order to facilitate debugging this CIT test failure in the future, this
change makes a few changes to ensure that whenever the script exits with
a non-zero error code, that there was something written to stderr.

The most likely cause of the script failure was that the CPU was powered off.

Previously there was a limitation that the CPU had to be powered on. This
limitation has already been addressed in the FPGA, so this PR will
also remove that restriction.

Testing performed:
1. With the CPU off, ran the oob_eeprom CIT test:
python3 /usr/local/bin/tests2/cit_runner.py --platform elbert -r tests.elbert.test_oob_eeprom

Repeated with the CPU on.

2. Ran oob-eeprom-util.sh dump, save, read, write, and program commands
Did this with the CPU off and repeated with the CPU on

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/143

Reviewed By: binhuang00

fbshipit-source-id: 8424ed5cfc