fscd: Fix test assert when parsing sensors
Summary:
Move the default attribute setting of offset and offset_table
to None to before the parts which can assert. This allows us
to be a bit imperfect in unit testing configurations but still
have the rest of the code work.

This fixes the unit tests throw exception that sensor data does
not have the attribute offset.

Test Plan: All unit tests pass.

Reviewed By: vineelasmile

fbshipit-source-id: ca67dc6d5f