gpio-cli: Ensure that __TEST__ is pased when compiling C objects
Summary:
Fix unit-test failures where the workarounds for the paths were
not in place due to missing __TEST__ defines in gpio_sysfs.c.

Test Plan:
1. ./tests2/ptest-runner libgpio-ctrl
(all tests pass)
2. build tiogapass and load on qemu. Ensure that the GPIOs
are exported correctly with the correct sysfs path (We did not
accidentally set the flag for prod version of the source) and
`gpiocli get-value --shadow LED_POST_CODE_0`.

Reviewed By: benwei13

fbshipit-source-id: 13279d6692