Attestation Util Interface Implementation
Summary:
This is an attestation util used for sending allowed payloads using
MCTP. Initially this will be used to send SPDM messages to the NIC via
MCTP, however it has beem designed with the addition of new payload
types and transport protocols in mind.

Because the current generation of NICs don't support SPDM over MCTP this
util is just an interface that will be used to develop tooling around
until the new hardware/firmware is available.

Test Plan:
**Unit Tests are in a separate diff for easier code review.**
 ---
Builds on Yv3
 ---

```
$: ./attest-util -h
MCTP Attestation Tool
Usage: ./attest-util [OPTIONS] SUBCOMMAND

Options:
  -h,--help                   Print this help message and exit
  -v,--version                Get version.

Subcommands:
  SPDM                        Sends an spdm message to the designated device and returns the response.
```

```
$: ./attest-util SPDM -h
Sends an spdm message to the designated device and returns the response.
Usage: ./attest-util SPDM [OPTIONS]

Options:
  -h,--help                   Print this help message and exit
  -e,--eid TEXT REQUIRED      Endpoint ID
  -f,--file TEXT:FILE REQUIRED
                              File name
  -o,--output TEXT:{json,raw} REQUIRED
                              Output type
```

```
$: ./attest-util SPDM -e 0 -f doesnt_exist -o json
--file: File does not exist: doesnt_exist
Run with --help for more information.
```

```
$: ./attest-util SPDM -e 0 -f invalid_spdm_payload -o json
{"payload":"DUMMYRESPONSE","status":"Payload is not an SPDM message","version":"0.1"}
```

```
$: ./attest-util SPDM -e 0 -f valid_spdm_payload -o json
{"payload":"DUMMYRESPONSE","status":"Success","version":"0.1"}
```

Reviewed By: amithash, benwei13

fbshipit-source-id: e28f10e786