pal: Add a dummy plat/meson.build to common
Summary:
This prevents the need for a platform layer to
have to provide a dummy plat/meson.build if they
do not have any overrides or even platform overrides
and rely solely on the common source.

Test Plan:
Build all the platforms.
Build angelslanding and check that the source in build-angelslanding
contain the correct plat/meson.build

Reviewed By: benwei13

fbshipit-source-id: 009f1e2378