rackmond: Record CRC and Timeout errors for all commands
Summary:
This expands the behavior to count the errors seen communicating
with a specific PSU for all commands. This is done by moving the
increment to the modbus_command() instead of done in the caller
context and only for a specific command. That way, we get an idea
of timeouts for individual calls, not just the the ones which fail even
after retries.

Test Plan:
Run stress test and record the now larger number of errors using
rackmonstatus

Reviewed By: benwei13

fbshipit-source-id: 504a1fbe07