qemux86: Remove perl from ptest rpdepends instead of removing the test
Summary:
This fixes issue where we have stopped ptest from being built but
we end up installing it anyway, so we just push the error further
down the build failure on dunfell.

Note, this does not mean the ptest would magically work. It will
still be broken since we are not installing perl.

Test Plan: Build qemux86 on dunfell

Reviewed By: benwei13

fbshipit-source-id: 2b60e3306b