Elbert management interface 1G support (#179)
Summary:
Change the fixed-speed to 1000 for both u-boot and the kernel.

For A2 support, specify mac3-clk-delay in both the U-boot and
kernel dts files. This config only is used in A2 and greater.

Enhance the u-boot logic to program the BCM53134 registers that
control the link speed. Program them to 1G.

Implement a setup_bcm53134.sh to program the BCM53134 registers
in the kernel.

With these changes, the default register values in the EEPROM
are overridden. Therefore, the newer BMC image will work with
older EEPROM images. However, older BMC images will not work
with newer EEPROM images.

Also note that if the u-boot image has the newer version,
but the kernel image is the older image, then the management
interface will not work. Therefore, the u-boot image should
not be programmed independently.

Testing performed:
1. Broke into u-boot and verified the management link worked:
ELBERT-boot=> setenv ipaddr 172.24.146.143
ELBERT-boot=> ping 172.24.146.129
ftgmac@1e690000: link up, 1000 Mbps full-duplex mac:c2:06:e4:8c:b3:b4
Using ftgmac@1e690000 device
host 172.24.146.129 is alive

2. Booted to the kernel. Inspected the /var/log/boot file
and verified the BCM53134 registers were not re-programmed.
Verified ssh to the BMC worked.

3. Programmed the u-boot image to be the existing 100M version
and the kernel image to be the new 1000M.

Booted to the kernel. Observed the following message in /var/log/boot:
BCM53134 registers not as expected. Configuring now.
Re-program the BCM53134 EEPROM to fix this permanently.

Verified ssh to the BMC worked.

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/179

Reviewed By: binhuang00

fbshipit-source-id: 3f416bdb38