linux-aspeed: add UBIFS config snippet
Summary:
Create a kernel config snippet enabled off the mtd-ubifs, which
enables and disables appropriate kconfig settings for UBIFS.  Any
system with the MACHINE_FEATURE 'mtd-ubifs' will gain these
settings automatically so that the machine kconfig does not need
to explicitly be modified.

Test Plan: Built and booted on qemu with Northdome.

Reviewed By: tao-ren, GoldenBug

fbshipit-source-id: f06f23a459