obmc-mctp lib returns wrong raw length
Summary:
When getting raw response from mctp traffic the total response is
returned, source eid + length byte + payload. However the length
returned is only the length of given by the length byte in the header
which does not include the source eid and the length itself.

When returning the raw packet we should return size + 2 to account for
those bytes.

Test Plan:
Original
- Raw packet: `00 0B 05 10 04 00 00 00 02 00 10 00 11`
- Size: `11`

Extended Size
- Raw Packet: `00 0b 05 10 04 00 00 00 02 00 10 00 11`
- Size `13`

Reviewed By: benwei13

fbshipit-source-id: 16b1ba6119