wedge100: setup_i2c.sh: delay pfe1100 creation if needed
Summary:
If the chassis is power cycled, i2c devices behind i2c-mux 7-0070 are not
reachable when setup_i2c.sh is executed, and this is because 7-0070 is not
powered until after power-on.sh.

As a result, we will skip pfe1100 device creation and fixup such devices
in sensor-setup script.

Test Plan:
Power cycled the chassis and verified pfe1100 devices are created after
bootup.

Reviewed By: binhuang00

fbshipit-source-id: 53566e7509