common: remove INHIBIT_PACKAGE_STRIP
Summary:
INHIBIT_PACKAGE_STRIP is a Yocto package setting that causes
it to not run 'strip' in the package during the install step.
This causes executables to remain quite large because they
contain all of the debug symbols.

I can't think of any reason to maintain this behavior and my
git-archaeology indicates that all usages of it are from a
copypasta at the time the recipe was first created.  Disabling
this setting yields a significant reduction in the flash image
size:

    30250624 flash-northdome-20210511184724
    29330946 flash-northdome-20210511210424

Test Plan: Build and boot in QEMU on Northdome.

Reviewed By: amithash

fbshipit-source-id: 174bcbbc38