rackmond: Correctly implement write-holding-reg as per specification
Summary:
The implementation of the current "write_holding_reg" assumes
that we can write multiple registers in one command using the
0x06 function. Unfortunately, this is not meeting the specification.
(https://en.wikipedia.org/wiki/Modbus#Function_code_6_(preset/write_single_holding_register))

The 0x06 function allows one to write a single word/register and not
multiple registers. We just happened to work because the only
entity using the "write_holding_reg" method is the one setting
the baudrate which just happens to write to a single register which
ends up meeting the specification for the 0x06 function.

The change acknowledges the fact that we can write only one register
and hence changes the prototype to make the function a lot more easy
to use (and also remove the need for dynamic allocation).

Test Plan:
Build rackmon package for wedge400. Flash on BMC with PSUs
supporting a baudrate of 115200.
set desired baudrate by editing /etc/sv/rackmond/run and adding
export RACKMOND_DESIRED_BAUDRATE=115200
and restart rackmond.
Observe rackmonstatus shifts to 115200 baud-rate and we can send and
receive modbuscommands:
root@rsw046-oob:~# modbuscmd -x 7 -t 500 a403007a0001
Response: a4 03 02 a0 0b cd 9a
(This proves that we were able to set the baudrate command, and the PSU
has switched to the new baudrate proving that the write-holding-reg
command worked)

Reviewed By: zhdaniel12

fbshipit-source-id: ca21ee1567