Increase the number of retries in init.d/hostname.sh
Summary:
[Common] Increase the number of retries in init.d/hostname.sh
In some DC, the existing 10 retries doesn't seem to be enough. So increased it to 20.
In order to make BMC keep retry for a longer time without adding too much load to our DNS,
we use exponental back-off. That is,
1) Up to retry iteration #10, wait 15 seconds in between retries (No Change)
2) After iteration #11 until #20 (end), doubles the wait time in every iteration
  (30,60,120,240, ...  with the max 3600 seconds)

Test Plan: Tested on fsw001.p047.f01.atn6 and fuji.01.07.mpk17

Reviewed By: amithash

fbshipit-source-id: 116428440e