Adds filtering by timestamp to log-util
Summary: This adds the option to both print and clear logs in a timestamp range to log-util. The CLI calls remain backwards compatible. When adding this in I added in some functionality to the sel format, including checking if it is in a timestamp range and creating a timestamp clear log that conforms to the standard log format.

Test Plan:
Manual testing. Please see this quip document for a demonstration of how to use the new functionality.

https://fb.quip.com/tysLAncdekKi

Additionally, tests added to selstream and log-util.

Reviewed By: amithash, benwei13

fbshipit-source-id: a2d35e471a