circleci: remove build/unit-test dependency
The 'unit-test' step is used to build and test under the
qemux86/ptest environment.  The 'build' step used to build
all of the various machine configs.  There is no need for a
dependency between these two steps.

When 'unit-test' fails, which it has been for a while, we end
up receiving no signal on the 'build' steps.  This greatly
reduces the utility of CircleCI.

Remove the dependency so we get signal on the various machine
build steps impacted by a PR.