ELBERT: PIM fault logging support (#191)
Summary:
Enhance the dpm_dump.sh script to dump and clear the faults
for PIM DPMs. Without this change, the fault logs on the PIM
DPMs fill up which means the most recent faults are not available.

A new pim_dpm_dump.sh script is introduced to
either dump or dump and clear faults for a specified PIM.
The existing dpm_dump.sh script then calls this new script
for each PIM.

Since the PIMs have UCD9090A, it is necessary to clear the
brownout bit in the MISC_CONFIG register before clearing faults
in order for the faults to be cleared from non-volatile store.

Also, the RUN_TIME_CLOCK(0xD7) command is run to set the time
so that the timestamps in the fault logs will be correct.

As part of this change, remove RUN_TIME_CLOCK(0xD7) and MISC_CONFIG(0xFC) commands
from the list of protected commands to allow fault clearing to work.

Fixed lint issues in board-utils.sh.

Testing performed:

1. After boot, verify both SWITCHCARD and PIM entries are present in /mnt/data1/log/dpm_log.
For example:
```
{'Date':'2021-06-28 21:47:29','Source':'SWITCHCARD','Log':'0x010200000000000000000000020200000000','detail_0':{'m':'0x04c0b281488049760000','e':'0x12'},'detail_1':{'m':'0x04c0b29a8d8049764725','e':'0x12'},'detail_2':{'m':'0x04c0b29a8d0049764125','e':'0x12'}}
{'Date':'2021-06-28 21:47:29','Source':'PIM2','Log':'0x010102020002020200000000','detail_0':{'m':'0x04c01bb288004976a228','e':'0x16'},'detail_1':{'m':'0x04c01bd989804976802f','e':'0x16'},'detail_2':{'m':'0x04c01be18a0049766c2f','e':'0x16'},'detail_3':{'m':'0x04c01be2480049760000','e':'0x16'},'detail_4':{'m':'0x04c01be88a8049766a36','e':'0x16'},'detail_5':{'m':'0x04c01bf488804976762f','e':'0x16'}}
```

2. Power cycle a linecard and verify that the clock settings are not lost by
reading the RUN_TIME_CLOCK(0xD7) command after power cycle.

3. Modify the dpm_dump.sh script to have a lower max_loglines value. Reboot
and verify that the dpm log was rotated.

4. Run the show_tech script in verbose mode and verify both dpm log files are
   present.

5. Performed a soft reboot and verified that dpm logs were logged and cleared.

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/191

Reviewed By: binhuang00

fbshipit-source-id: 39c61a5181