circleci: remove build/unit-test dependency
Summary:
The 'unit-test' step is used to build and test under the
qemux86/ptest environment.  The 'build' step used to build
all of the various machine configs.  There is no need for a
dependency between these two steps.

When 'unit-test' fails, which it has been for a while, we end
up receiving no signal on the 'build' steps.  This greatly
reduces the utility of CircleCI.

Remove the dependency so we get signal on the various machine
build steps impacted by a PR.

Test Plan:
Create PR on Github and confirm expected behavior.
    - https://app.circleci.com/pipelines/github/facebook/openbmc/1898/workflows/558670fe-7f62-4a4c-a00c-cb1639c98b9c

Reviewed By: jerrt2003

fbshipit-source-id: 3a2c8bb8d7