optional bic gpio injection through false file path
Summary: In order to read the gpio from the BIC, we normally would send an IPMI command to get the response. However, this adds in the option to add a file in the system, and if that file exists, it will override the value received from the bic with the value in the file.

Test Plan: Manual testing.

Reviewed By: benwei13

fbshipit-source-id: 34bf9a5ca8