added check for config file existence
Summary: This will check to see if there is a config file, and return with helpful message if there is not one present. This will replace the confusing json parse error when there is not a config file present.

Test Plan:
Manual Testing. Please see this paste for output:
P429424384

Reviewed By: GoldenBug

fbshipit-source-id: d63006e3f6