added ability to inject values
Summary:
If compiled with the injection flag, this will check to see if there is a false gpio value file to read from before it reads from the standard gpio file. If there is a file, it will read from that, otherwise it will read from the standard gpio file as always.

If not compiled with the flag, functionality is unchanged.

Test Plan:
Manual testing.
1. Check to make sure can read without false path
2. Add false path and check that it reads a new (different value)
3. Change value in false path to make sure it can read either 0 or 1

See this paste: P430453159

Reviewed By: benwei13

fbshipit-source-id: 3d54631890