additional error checking for bic gpio injection
Summary:
This will add in the `tmp/gpio` directory if it does not exist already. Its absence causes a failure in the file creation step, and prior to this there was no clear error to message as to why it failed.

This fix is already in the regular gpio, as I discovered it while developing that, then added it to this after.

Test Plan:
Manual testing.
1. deleted the `/tmp/gpio` dir
2. ran the utility
3. printed the logs to ensure that utility ran correctly
4. ran utility again to ensure that it could run whether or not the directory was present

Reviewed By: benwei13

fbshipit-source-id: 483ac10533