ELBERT: updates FSCD invalid sensor retries for hitless bios_util.sh (#203)
Summary:
### Problem
FSCD is currently shutting down the x86 CPU during bios_util.sh because sensord doesn't poll the sensors while that utility is running. The reason we paused the polling is because we were seeing instability with bios_util.sh and found that pausing the sensor polling solved that issue.

### Solution
This change bumps the "invalid" threshold in the Elbert FSCD configs. This threshold is the number of invalid sensor reads that will be tolerated before triggering the related action. Since the FSCD polling interval is 2 seconds, I've set the threshold for invalid reads to 60, which is equivalent to 3 minutes. I benchmarked the bios_util.sh commands and found that they typically take ~1m10s but can take over 2m, so I set the threshold to 3 minutes to provide a bit of buffer.

## Testing
On Elbert, verified that the x86 CPU is not shut down during bios_util.sh and that FSCD is not failing.

Also upgraded our regression testbed to verify x86 status after running bios_util.sh; we have been testing this for the past week and seen no issues.

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/203

Reviewed By: binhuang00

fbshipit-source-id: ee346caf0e