Add YosemiteV3 syseventgen.conf layer
Summary: Adding a copy of syseventgen.conf to the Yv3 image for use by syseventgen-util. It's literally just a copy of the default config file from the core recipe. I tried just extending the existing .bb recipe to add a config file without touching the executable, but that didn't work, so I included binary installation in the Yv3 recipe.

Test Plan:
First I just verified that the current image _doesn't_ include a `syseventgen.conf` file by default:

  root@sled3103-oob:~# syseventgen-util
  Configuration file needed.
  Please create one and place it at /etc/syseventgen.conf

  root@sled3103-oob:~# ls -l /etc/syseventgen.conf
  ls: /etc/syseventgen.conf: No such file or directory

Then I installed a signed image with the changes and rebooted and verified this message didn't get printed, and that the configuration file was present at `/etc/syseventgen.conf`:

  root@sled3103-oob:~# syseventgen-util
  System Event Generation Util
  Usage: syseventgen-util [OPTIONS] [SUBCOMMAND]

  Options:
    -h,--help
    -v,--verbose                Turn on verbose outputs

  Subcommands:
    clear                       Get rid of all injected SEL's
    print                       Print all of the injected SEL blocks
    0                           set_sensor_slot1
    1                           set_sensor_slot2
    2                           no threshold
    3                           ipmi oem_ras_sel
    4                           ipmi sensor alert
    5                           ipmi storage add sel
    6                           ipmi sensor plat 7-bytes
    7                           ipmi sensor plat 8-bytes
    8                           payload
    9                           bic slot 1 #27

  root@sled3103-oob:~# ls /usr/local/bin/syseventgen-util
  /usr/local/bin/syseventgen-util

  root@sled3103-oob:~# ls /etc/syseventgen.conf
  /etc/syseventgen.conf

Reviewed By: benwei13

fbshipit-source-id: ab2a3121ac