rocko: fix a few more python packages
Summary:
We would like to upgrade aiohttp and some of the dependent packages
but that is relatively unsafe to do on rocko at this point.  Add
all the python packages used by the rest server into the rocko pin
list.

Test Plan:
Build tiogapass and check manifest before and after the (local) aiohttp
change to confirm no packages change.

Reviewed By: GoldenBug

fbshipit-source-id: eda9108bed