sonorapass: remove gdb
Summary:
Sonorapass is a shelved platform but we currently keep it around for
build tests.  Since it has a smaller flash layout we are starting to
run out of space.  Remove gdb from it to free up a good amount of
space and to prevent build failures due to "FIT image too big" errors.

Test Plan: Build sonorapass and confirm gdb is no longer in manifest.

Reviewed By: GoldenBug

fbshipit-source-id: d4243a41fa