zeus: tpm2-tools: remove useless bbappend
Summary:
When building for a dunfell machine I noticed the following warning:
```
  WARNING: No recipes available for:
    ./yocto/lf-dunfell/poky/../../../common/recipes-tpm2/tpm2-tools/tpm2-tools_3.2.0.bbappend
```

This was due to an unused bbappend in common.  Investigating further
this seems to have originally been used for Zeus but now Zeus has a
future version of the tpm2-tools package in backports.  Remove the
unused bbappend.

```
  $ find -name tpm2-tools_3.2.0.bb
  ./yocto/zeus/meta-security/meta-tpm/recipes-tpm2/tpm2-tools/tpm2-tools_3.2.0.bb
  $ ls meta-facebook/meta-backports/zeus/recipes-tpm2/tpm2-tools
  tpm2-tools_4.1.1.bb
```

Test Plan:
Zeus based machines build in CIT and warning is gone for
others.

Reviewed By: amithash

fbshipit-source-id: 564fbab8f6