openbmc-init: fix fby35 distro choice
Summary:
It was reported that a reorder of the fby35 system after fby3
caused fby35 to be built with dunfell instead of zeus.  Fix
the logic to ensure that the meta-layer selection is an exact
match with the set in PLAT_DISTRO_OVERRIDES rather than any partial
match.

The code already handles sub-machine types because it searches by
containing meta rather than by machine-name.

Fixes facebookexternal/openbmc.wiwynn#1732.

Test Plan:
Confirm the following distro selections:

    fby35 - zeus
    fby3 - dunfell
    fby2 - rocko
    fby2-ubu - rocko
    wedge400 - dunfell
    wedge - rocko
    wedge100 - rocko

Reviewed By: amithash

fbshipit-source-id: 177858c894