sync_yocto: enable secondary repo patches
Summary:
We are going to need to backport changes to Yocto (rocko and zeus) that
are not in the current upstream branches, and since those branches are
EOL will likely never have these patches.  Adding support to the github
`sync_yocto` script to be able to apply patches into the public side of
our Yocto repositories.

Test Plan:
Manually applied into a clean clone of github.com/facebook/openbmc and
confirmed that patches applied into Yocto clones.

Reviewed By: GoldenBug

fbshipit-source-id: 81246aabb5