ELBERT: remove redundant fan LED logic from led-controld (#220)
Summary:
Currently, both FSCD and led-controld are updating the individual fan LEDs. This change removes the fan LED control logic from led-controld (not the fan status LED though, that is still controlled by led-controld), as it is redundant to the FSCD LED logic.

## Testing
Tested on Elbert. Verified that fan LEDs are correct on boot and that fan status LED updates correctly based on fan statuses.

Pull Request resolved: https://github.com/facebookexternal/openbmc.arista/pull/220

Reviewed By: tao-ren

fbshipit-source-id: ab725891a8