misc-utils: Add biview helper C++ container
Summary:
This is a container class which privides
a bi-directly read-only (const) map.

There are a lot of places where we would want to
do bi-directional look-up. Example, given sensor-id,
find a sensor name and vice-versa. A container for
such a data-store will be really helpful.

Test Plan: Build into test application.

Reviewed By: GoldenBug

fbshipit-source-id: 10579d7924