Add Redfish ComputerSystem handler
Summary:
Adding `ComputerSystem`, which is a parent for `Bios`, which is a parent for `FirmwareDumps` (introduced in the next Diffs). Current implementation exists only to provide link to `Bios`, which exists only to provide link to `FirmwareDumps`, so other fields are omitted at this time.

See examples of `ComputerSystem` in https://www.dmtf.org/sites/default/files/standards/documents/DSP0266_1.12.0.pdf

Test Plan: See Test Plan in: D31400585

Reviewed By: deathowl

fbshipit-source-id: 8617d9d8e4