fruid: fix buffer-overrun
Summary:
It was reported on Northdome that certain invocations of `fruid-util`
could result in an error such as:
```
    malloc(): invalid size(unsorted)
```
Looking at the glibc source code this appears to be a buffer-overrun
resulting in corruption of the buddy information in the heap allocation.

I visually inspected the code and found two instances where strings were
not properly NUL-terminated.  I then hacked at the executable to get it
to compile on an x86-64 system so I could run the tool against an EEPROM
capture that was reported to cause the problem.  I enabled GCC's ASAN
and it identified the following problem:

```
$ gcc recipes-lib/fruid/files/fruid.c recipes-core/fruid/files/fruid-util.c \
    -Irecipes-lib/fruid/files -Irecipes-lib/ipmi/files \
    -Irecipes-lib/obmc-pal/files -ljansson -O1 -g -fsanitize=address

$ ./a.out slot1 --modify --CCD1 9KF8120U10147 5.bin
=================================================================
==166643==ERROR: AddressSanitizer: heap-buffer-overflow on address 0x6020000003cf at pc 0x55c4c24403db bp 0x7ffc2cb11cf0 sp 0x7ffc2cb11ce0
READ of size 1 at 0x6020000003cf thread T0
    #0 0x55c4c24403da in rtrim recipes-core/fruid/files/fruid-util.c:70
    #1 0x55c4c2440e7d in fruid_trim recipes-core/fruid/files/fruid-util.c:87
    #2 0x55c4c2440e7d in get_fruid_info recipes-core/fruid/files/fruid-util.c:511
    #3 0x55c4c2444ad7 in do_action recipes-core/fruid/files/fruid-util.c:1148
    #4 0x55c4c2444fe3 in main recipes-core/fruid/files/fruid-util.c:1273
    #5 0x7fd8bd86b7fc in __libc_start_main (/lib64/libc.so.6+0x237fc)
    #6 0x55c4c2436459 in _start (/home/apwillia/local/sync/fbopenbmc/common/a.out+0x7459)

0x6020000003cf is located 1 bytes to the left of 2-byte region [0x6020000003d0,0x6020000003d2)
allocated by thread T0 here:
    #0 0x7fd8bdacd777 in __interceptor_malloc (/usr/lib/gcc/x86_64-pc-linux-gnu/11.2.0/libasan.so.6+0xb6777)
    #1 0x55c4c2436b4f in _fruid_area_field_read recipes-lib/fruid/files/fruid.c:192

AddressSanitizer: heap-buffer-overflow recipes-core/fruid/files/fruid-util.c:70 in rtrim
```

Through some additional debug I determined that the `chassis.custom2`
field in this EEPROM was initialized with 0's.  This causes
`_fruid_area_field_read` to interpret the data as {TYPE_BINARY, len 0}
for which it allocates a 1-byte NUL-terminated string.  Later on the
`rtrim` function does not correctly handle strings where `strlen` is 0
without causing corruption past the end of the string.  Added a case to
skip `rtrim` when the string is 0-bytes long.

After this was resolved, ASAN identified another problem:
```
$ ./a.out slot1 --modify --CCD1 9KF8120U10147 5.bin
=================================================================
==170880==ERROR: AddressSanitizer: heap-buffer-overflow on address 0x611000000110 at pc 0x7fa7c0d8481b bp 0x7fffe02a33c0 sp 0x7fffe02a2b68
WRITE of size 14 at 0x611000000110 thread T0
    #0 0x7fa7c0d8481a in __interceptor_memcpy (/usr/lib/gcc/x86_64-pc-linux-gnu/11.2.0/libasan.so.6+0x3a81a)
    #1 0x55dd7892aced in memcpy /usr/include/bits/string_fortified.h:29
    #2 0x55dd7892aced in fruid_modify recipes-lib/fruid/files/fruid.c:1852
    #3 0x55dd7892fb2c in do_action recipes-core/fruid/files/fruid-util.c:1148
    #4 0x55dd78930062 in main recipes-core/fruid/files/fruid-util.c:1278
    #5 0x7fa7c0b9e7fc in __libc_start_main (/lib64/libc.so.6+0x237fc)
    #6 0x55dd78921459 in _start (/home/apwillia/local/sync/fbopenbmc/common/a.out+0x7459)

0x611000000110 is located 0 bytes to the right of 208-byte region [0x611000000040,0x611000000110)
allocated by thread T0 here:
    #0 0x7fa7c0e00777 in __interceptor_malloc (/usr/lib/gcc/x86_64-pc-linux-gnu/11.2.0/libasan.so.6+0xb6777)
    #1 0x55dd78927e9d in fruid_modify recipes-lib/fruid/files/fruid.c:1618
```

The source of this issue is found when adding a new optional field, such
as `CCD1`, it can cause an entirely new section (ex. `chassis`) to be
added with all of its optional fields becoming 1-byte type/len entries.
In order to prevent a buffer overrun when creating the modified EEPROM
image, we need to account for these additional optional fields.
Over-estimate this size by adding an allocation for all optional fields
in the entire image.

Test Plan: Fix issues and confirm ASAN no longer reports problems.

Reviewed By: amithash

fbshipit-source-id: 6729ce2a43